[Refactor] Extract litegraph settings sync to a composable
